refactor: security -> config 패키지 명 변경
